% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npboy.R
\name{npboy}
\alias{npboy}
\title{npboy}
\usage{
npboy(a, b, c, miu, sigma)
}
\arguments{
\item{a}{Newspaper retail price}

\item{b}{Newspaper purchase price}

\item{c}{Newspaper rebate(a>b>c)}

\item{miu}{The expected number of daily newspaper buyers}

\item{sigma}{The variance of the number of daily newspaper buyers}
}
\value{
The optimum number of newspapers a newspaper boy should buy each day \code{n}
}
\description{
A function to solve the newspaper-boy-problem in mathematical modeling.
}
