% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_add}
\alias{rcpp_add}
\title{rcpp_add}
\usage{
rcpp_add(x, y)
}
\arguments{
\item{the}{second parameter y (y)}
}
\value{
the sum of two integers \code{n}
}
\description{
The sum of two integers.This is a simple function to test the integrity of Rcpp package
}
\examples{
\dontrun{
x<-3
y<-4
z<-rcpp_add(3,4)
}
}
