\name{SC19017016-package}
\alias{SC19017016-package}
\docType{package}
\title{
  Solve simple problems in mathematical modeling
}
\description{
  After building the mathematical model, we usually need the corresponding code to implement it.This package include some simple and common model,such as the newspaper boy problem and the waste of steel rolling probelm.The function will be provide to help a beginner to learn the Mathematical modeling course. 
}

\author{
Zifeng Ye, zfyeee@126.com.

Maintainer: Zifeng Ye <zfyeee@126.com>
}

\keyword{ package }

