% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steel.R
\name{steel}
\alias{steel}
\title{Waste in steel rolling}
\usage{
steel(L, sigma, pre)
}
\arguments{
\item{L}{Required steel's size}

\item{sigma}{Accuracy of instrument}
}
\value{
The mean of finished rolled steel should be set \code{l}
}
\description{
A function to solve the problem about wasting in steel rolling
}
